refactor: move entry point to application-harness directory
