+++ packages/spectacular/src/lib/application-testing/application-harness/bootstrap-spectacular-application.spec.ts
